update spi
the last two partitions are propably bs as the spi chips is only 256kb
long and the last two partitions start beyond 256k. Also nvram and
kernel are dubious to be on a spi flash chip. (especially the kernel).
Most likely scenario is that this partition table is just incorrect.
Will do a dump of the whole chip and then figure out the partition table
manually.