remove #size-cells #address-cells from nand (59)
Maybe this will fix the partitions not showing up, as they are child
nodes of this node, and these two properties affect child nodes only.

now, when a cable is plugged in, both leds of the network port light up.

Also all nine leds driven by stp are solid on.