dts: update syscon-reboot
Remove regmap comment. the `regmap` property is deprecated.
Instead of putting the reboot node at the top of the tree and linking
to the registers by specifying a phandle (to the rcu in this case),
we should have the reboot node as a child of the rcu.

Also should look into reboot-modes some time, as the draytek bootloader
does communicate with DrayOS by passing magic values into some registers
that seem to be preserved across reboot.