lantiq: add draytek preloader
The DrayBoot bootloader loads the kernel from a hardcoded address
into ram (according to an image header, which specifies the size).
Then it copies the first 512K (after the header) to 0x80024000 and
jumps to it. This requires us to have some piece of code that fits in
512K that can decompress the kernel and jump to it. We can use
lzma-loader for this. Therefore it has been modified to allow searching
for the OKLI uImage from arbitrary memory addresses instead of only
memory mapped nor flash at KSEG1.

Note that the DrayTek image header still needs to be implemented in
firmware-tools, so this is not ready for flashing yet.