lantiq: drop KERNEL_ENTRY
KERNEL_ENTRY should only be used it it differs from KERNEL_LOADADDR.
In our case it does not, so specifiying it adds nothing. When we change
KERNEL_LOADADDR (in the case of draytek devices) we also need to change
KERNEL_ENTRY, which can easily be forgotten and should be unnecessary.