kernel: add options for vsockets and virtio ports
Add menuconfig options to include modules for virtio vsockets. These
are used when running as Guest OS in a VM.

OpenWRT can be run in a vm (mostly relevant to the x86/64 and armsr targets).
Often it is convenient to have some sort of guest agent running, to ease
communication from host to guest.
Virtual Sockets provide this communication channel.
Virtio is a transport for this communication channel.

Virtual Sockets over virtio are for example used by Incus.
Virtual Sockets in general are used by most hypervisors, including VMware,
Hyper-V, and libvirt (qemu). These may need other transport-specific
modules not included in this commit.

Note that on the x86/64 and loongarch64 targets the `vsock` and
`vmw_vsock_virtio_transport_common` modules are built into the kernel
(CONFIG_VSOCKETS and CONFIG_VIRTIO_VSOCKETS_COMMON are set),
so on those targets they are not a dependency of the other modules.