kernel: fix vsock kmods
Don't build kmod-vsock and kmod-vsock-virtio-common on x86_64 and
loongarch64, as these modules are already built into the kernel.
There is no point in having them as modules too.

Fix dependencies:
- CONFIG_VIRTIO_VSOCKETS depends on CONFIG_VIRTIO (which is
  indicated by VIRTIO_SUPPORT in openwrt)

I am not sure what the best approach is here. Should we not have the
kmod packages at all if the modules are built into the kernel, as they
wouldn't be loaded anyway, or is that going to cause only more
confusion?