kernel: vsock kmods: drop target specific DEPENDS
It turns out the OpenWRT kmod packaging system already deals with
all of this...

If a module is built-in, the system just skips it and ships an empty
package only containing the file in modules-load.d (if it has it).

That is totally fine as it takes up only literal bytes, not even kbytes
and it massively simplifies the dependency situation. If the module
is all of a sudden not built into the kernel anymore because the
config changes, it will now automatically built as a module and contain
the module file if it gets installed, so the net effect is identical.
Way better than what I was trying to do.